requirements.txt
Django==1.9.5
Pillow==3.2.0
PySocks==1.5.7
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
httplib2==0.9.2
markdown2==2.3.1
pytz==2016.6.1
six==1.10.0
twilio==5.4.0
